requirements.txt
numpy==1.19.1
oauthlib==3.1.0
requests==2.24.0
python-telegram-bot==12.8
pylatexenc==2.7
sympy==1.6.2
more-itertools==8.5.0
